requirements.txt
asgiref==3.10.0
asttokens==3.0.0
boto3==1.35.76
botocore==1.35.99
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coverage==7.6.9
decorator==5.2.1
dj-database-url==2.3.0
Django==5.1.7
django-configurations==2.5.1
django-cors-headers==4.9.0
django-filter==24.3
django-model-utils==5.0.0
django-storages==1.14.4
django-unique-upload==0.2.1
djangorestframework==3.15.2
executing==2.2.1
factory_boy==3.3.1
Faker==37.11.0
flake8==7.1.1
ghp-import==2.1.0
gunicorn==23.0.0
idna==3.11
iniconfig==2.3.0
ipdb==0.13.13
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
Markdown==3.9
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mock==5.1.0
newrelic==10.3.1
openrouteservice==2.3.3
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyyaml_env_tag==1.1
requests==2.32.5
s3transfer==0.10.4
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
